Updated test cases to use mocked values, split out test classes and used toIntOrNull() extension function instead of try/catch
