+++ app/src/main/kotlin/com/waz/zclient/core/network/pinning/CertificatePin.kt
@@ -4,16 +4,11 @@ data class CertificatePin(
     val domain: String,
     val certificate: ByteArray
 ) {
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (javaClass != other?.javaClass) return false
-
-        other as CertificatePin
-
-        if (domain != other.domain) return false
-        if (!certificate.contentEquals(other.certificate)) return false
-        return true
-    }
+    override fun equals(other: Any?) =
+        this === other
+            && this.javaClass == other.javaClass
+            && domain == other.domain
+            && certificate.contentEquals(other.certificate)
 
     override fun hashCode(): Int {
         var result = domain.hashCode()
